+++ lib/darrrr/constants.rb
