+++ lib/darrrr/crypto_helper.rb
