+++ lib/darrrr/cryptors/default/default_encryptor.rb
