+++ lib/darrrr/cryptors/default/encrypted_data.rb
