+++ lib/darrrr/cryptors/default/encrypted_data_io.rb
