+++ lib/darrrr/recovery_provider.rb
