+++ lib/darrrr/recovery_token.rb
