+++ lib/darrrr/serialization/recovery_token_reader.rb
