+++ lib/darrrr/serialization/recovery_token_writer.rb
