+++ spec/lib/darrrr/account_provider_spec.rb
