+++ spec/lib/darrrr/cryptors/default/encrypted_data_spec.rb
