+++ spec/lib/darrrr/recovery_provider_spec.rb
