+++ spec/lib/darrrr/recovery_token_spec.rb
