+++ spec/lib/darrrr_spec.rb
